# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão e Monitoramento de Casos de Cólera",
        "description": "Documentação da API para Sistema de Gestão e Monitoramento de Casos de Cólera em Angola",
        "contact": {
            "name": "Suporte Técnico",
            "email": "admin@saudeangola.gov"
        },
        "license": {
            "name": "Licença MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Versão 1"
        }
    ],
    "paths": {
...
        "/unidades/{id}": {
            "get": {
                "tags": [
                    "Unidades de Saúde"
                ],
                "summary": "Obter detalhes de unidade de saúde",
                "description": "Retorna os dados de uma unidade de saúde específica pelo ID",
                "operationId": "obterUnidadeSaude",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade de saúde",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade de saúde encontrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade de saúde encontrada com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UnidadeSaude"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unidade de saúde não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade de saúde não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Unidades de Saúde"
                ],
                "summary": "Atualizar unidade de saúde",
                "description": "Atualiza os dados de uma unidade de saúde existente",
                "operationId": "atualizarUnidadeSaude",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade de saúde",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gabinete_provincial_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Centro de Saúde Municipal - Atualizado"
                                    },
                                    "diretor_medico": {
                                        "type": "string",
                                        "example": "Dra. Teresa Gomes"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Centro_Saude"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Av. Nova, 78"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "+244 923456790"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "centro.novo@saude.gov.ao"
                                    },
                                    "capacidade": {
                                        "type": "integer",
                                        "example": 75
                                    },
                                    "tem_isolamento": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unidade de saúde atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade de saúde atualizada com sucesso."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UnidadeSaude"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Unidade de saúde não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unidades de Saúde"
                ],
                "summary": "Eliminar unidade de saúde",
                "description": "Remove uma unidade de saúde do sistema",
                "operationId": "eliminarUnidadeSaude",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da unidade de saúde",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade de saúde eliminada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade de saúde eliminada com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Unidade de saúde não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnidadeSaude": {
                "title": "Unidade de Saúde",
                "description": "Modelo de Unidade de Saúde",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "gabinete_provincial_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 2
                    },
                    "nome": {
                        "type": "string",
                        "example": "Hospital Provincial de Luanda"
                    },
                    "diretor_medico": {
                        "type": "string",
                        "example": "Dra. Ana Silva"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Hospital_Geral"
                    },
                    "endereco": {
                        "type": "string",
                        "example": "Av. Principal, 123"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "+244 923456789"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "hospital@saude.gov.ao"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -8.839
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 13.289
                    },
                    "capacidade": {
                        "type": "integer",
                        "example": 200
                    },
                    "tem_isolamento": {
                        "type": "boolean",
                        "example": true
                    },
                    "capacidade_isolamento": {
                        "type": "integer",
                        "example": 20
                    },
                    "casos_ativos": {
                        "type": "integer",
                        "example": 45
                    },
                    "leitos_ocupados": {
                        "type": "integer",
                        "example": 120
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}